adjust routerFilter logic
