+++ spring-cloud-huawei-router/src/main/java/com/huaweicloud/router/client/loadbalancer/webmvc/WebMvcConfiguration.java
@@ -1,0 +1,38 @@
+/*
+
+ * Copyright (C) 2020-2022 Huawei Technologies Co., Ltd. All rights reserved.
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.huaweicloud.router.client.loadbalancer.webmvc;
+
+import org.apache.servicecomb.router.RouterFilter;
+import org.apache.servicecomb.router.distribute.AbstractRouterDistributor;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConditionalOnWebApplication(type = Type.SERVLET)
+public class WebMvcConfiguration {
+  @Bean
+  @ConditionalOnMissingBean(WebMvcServiceInstanceFilter.class)
+  public WebMvcServiceInstanceFilter canaryServiceInstanceFilter(
+      AbstractRouterDistributor<ServiceInstance> routerDistributor, RouterFilter routerFilter) {
+    return new WebMvcServiceInstanceFilter(routerDistributor, routerFilter);
+  }
+}
