+++ spring-cloud-huawei-router/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -17,3 +17,5 @@
 ## ---------------------------------------------------------------------------
 
 com.huaweicloud.router.client.RouterClientAutoConfiguration
+com.huaweicloud.router.client.loadbalancer.webflux.WebFluxConfiguration
+com.huaweicloud.router.client.loadbalancer.webmvc.WebMvcConfiguration
